+++ trl/__init__.py
@@ -93,6 +93,7 @@ _import_structure = {
         "RLOOTrainer",
         "SFTConfig",
         "SFTTrainer",
+        "DataCollatorForLanguageModeling",
         "DataCollatorForMTPLanguageModeling",
         "WinRateCallback",
         "XPOConfig",
@@ -192,6 +193,7 @@ if TYPE_CHECKING:
         RLOOTrainer,
         SFTConfig,
         SFTTrainer,
+        DataCollatorForLanguageModeling,
         DataCollatorForMTPLanguageModeling,
         WinRateCallback,
         XPOConfig,
