+++ trl/trainer/__init__.py
@@ -70,6 +70,7 @@ _import_structure = {
     "rloo_trainer": ["RLOOTrainer"],
     "sft_config": ["SFTConfig"],
     "sft_trainer": ["SFTTrainer"],
+    "data_collator": ["DataCollatorForLanguageModeling"],
     "mtp_data_collator": ["DataCollatorForMTPLanguageModeling"],
     "utils": [
         "RunningMoments",
@@ -141,6 +142,7 @@ if TYPE_CHECKING:
     from .rloo_trainer import RLOOTrainer
     from .sft_config import SFTConfig
     from .sft_trainer import SFTTrainer
+    from .data_collator import DataCollatorForLanguageModeling
     from .mtp_data_collator import DataCollatorForMTPLanguageModeling
     from .utils import (
         RunningMoments,
